Refactor filter imports to use new constants and utils structure
