+++ apps/ff-site/src/app/_components/LocationFilter.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { DEFAULT_LOCATION_FILTER_OPTION } from '@filecoin-foundation/hooks/useFilter'
+import { DEFAULT_LOCATION_FILTER_OPTION } from '@filecoin-foundation/hooks/useFilter/constants'
 import { FilterListbox } from '@filecoin-foundation/ui/FilterListbox'
 import type { OptionType } from '@filecoin-foundation/ui/Listbox/ListboxOption'
 import { LOCATION_KEY } from '@filecoin-foundation/utils/constants/urlParamsConstants'
