+++ apps/ff-site/src/app/blog/components/BlogContent.tsx
@@ -3,7 +3,8 @@
 import { useSearchParams } from 'next/navigation'
 
 import { useFilter } from '@filecoin-foundation/hooks/useFilter'
-import { DEFAULT_CATEGORY_FILTER_OPTION } from '@filecoin-foundation/hooks/useFilter'
+import { DEFAULT_CATEGORY_FILTER_OPTION } from '@filecoin-foundation/hooks/useFilter/constants'
+import { entryMatchesCategoryQuery } from '@filecoin-foundation/hooks/useFilter/utils'
 import { CardGrid } from '@filecoin-foundation/ui/CardGrid'
 import { NoSearchResultsMessage } from '@filecoin-foundation/ui/NoSearchResultsMessage'
 import { Pagination, usePagination } from '@filecoin-foundation/ui/Pagination'
@@ -24,7 +25,6 @@ import { BASE_DOMAIN } from '@/constants/siteMetadata'
 
 import { graphicsData } from '@/data/graphicsData'
 
-import { entryMatchesCategoryQuery } from '@/utils/filterUtils'
 import { getCategoryLabel } from '@/utils/getCategoryLabel'
 import { getSortOptions } from '@/utils/getSortOptions'
