+++ apps/ff-site/src/app/events/utils/filterUtils.ts
@@ -1,23 +1,8 @@
 import { VIRTUAL_EVENT_FILTER_OPTION } from '@/events/constants/constants'
 import type { Location } from '@/events/schemas/LocationSchema'
 
-type WithCategory = {
-  category: string
-}
-
 type WithLocation = {
   location: Location
-}
-
-export function entryMatchesCategoryQuery<Entry extends WithCategory>(
-  entry: Entry,
-  query?: string | Array<string>,
-) {
-  if (Array.isArray(query)) {
-    return query.includes(entry.category)
-  }
-
-  return entry.category === query
 }
 
 export function entryMatchesLocationQuery<Entry extends WithLocation>(
