+++ packages/hooks/src/useFilter/constants/index.ts
@@ -1,0 +1,5 @@
+export {
+  DEFAULT_FILTER_ID,
+  DEFAULT_CATEGORY_FILTER_OPTION,
+  DEFAULT_LOCATION_FILTER_OPTION,
+} from './filterConstants'
