+++ packages/hooks/src/useFilter/index.ts
@@ -1,6 +1,1 @@
-export {
-  DEFAULT_CATEGORY_FILTER_OPTION,
-  DEFAULT_FILTER_ID,
-  DEFAULT_LOCATION_FILTER_OPTION,
-} from './filterConstants'
 export { useFilter } from './useFilter'
