+++ packages/hooks/src/useFilter/utils/index.ts
@@ -1,0 +1,1 @@
+export { entryMatchesCategoryQuery } from './filterUtils'
